requirements.txt
numpy==1.19.2
click==7.1.2
eventlet==0.29.0
Flask==1.1.2
opencv-python==3.4.2.17
python-engineio==3.13.2
python-socketio==4.6.0
six==1.15.0
torch==1.6.0
tqdm==4.50.2
dnspython==1.16.0
future==0.18.2
greenlet==0.4.17
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
monotonic==1.5
Pillow==7.2.0
Werkzeug==1.0.1
